+++ SystemPayments/src/syspaym/utils/Stat.java
@@ -4,12 +4,12 @@ package syspaym.utils;
  * Created by Admin on 26.04.2016.
  */
 public class Stat{
-    public Float SumOfPayments;
+    public Double SumOfPayments;
     public Integer NumberOfPayments;
 
     public Stat()
     {
-        SumOfPayments = 0.0f;
+        SumOfPayments = 0.0D;
         NumberOfPayments = 0;
     }
 }
